deep-pulse/requirements.txt
# Core
cryptography>=42.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
# py-libp2p is unavailable/incompatible, disabled for Phase 1
feedparser>=6.0.11
pydantic

# Intelligence Orchestration
langchain>=0.1.0
langgraph>=0.0.10
langchain-openai>=0.0.1
langchain-community>=0.0.10

# Intelligence Orchestration
langchain>=0.1.0
langgraph>=0.0.10
langchain-openai>=0.0.1
langchain-community>=0.0.10

# Intelligence Orchestration
langchain>=0.1.0
langgraph>=0.0.10
langchain-openai>=0.0.1
langchain-community>=0.0.10

# Scouting & AI
crawl4ai>=0.3.0
crawl4ai[all]
playwright
sentence-transformers>=2.7.0  # For RLM Memory

# CLI & Utils
pyyaml>=6.0.1
